requirements.txt
pydantic
deepmerge
PyYAML
colorama